Refactor download stage into focused modules
Decompose 1722-line download.rs into single-responsibility modules:

  stages/download/
  ├── mod.rs      (316 lines) - Stage orchestration, public API
  ├── chunk.rs    (529 lines) - Chunk download with cache/retry/cancellation
  ├── bounded.rs  (376 lines) - Two-phase permit-bounded algorithm
  └── tests.rs    (591 lines) - Test infrastructure and cases

Benefits:
- Each module under 600 lines with clear purpose
- Bounded algorithm independently testable
- Chunk download logic isolated from stage orchestration
- Extracted spawn_cache_write helper for code reuse

No functional changes - all 1212 tests pass.
